refactor: dedupe usage accounting, isolate bash policy, harden rejection handler
Addresses maintainability findings from a code review:

- usage: new src/usage.ts owns the token helpers plus a PhaseUsage accumulator
  (per-session dedup, "cumulative total wins over deltas", aggregation). Shared by
  metadata, tui, and runner in place of three drifting copies of the same logic.
- bash-policy: move the allow/deny pattern lists and bashPolicy into
  src/bash-policy.ts; agents.ts now only consumes the assembled policy.
- main: swallow only the known-benign SSE abort (isIgnorableRejection); every other
  unhandled rejection now logs loudly with its stack (log.error, bypasses the TUI
  mute) instead of being hidden, without crashing a long-running run.
- tui-theme: add a fallback test for the opentui-internals adapter.

No behavior change in usage accounting or the bash policy. Tests: 122 pass.
